geoserver-docker/requirements.txt
# requirements.txt
# --------------------
ecoshard==0.4.0
Flask==1.1.2
GDAL==3.0.4
pygeoprocessing==2.0.0
requests==2.23.0
retrying==1.3.3
waitress==1.4.4
Cython==0.29.19
matplotlib==3.2.1
retrying==1.3.3
Rtree==0.9.4
scipy==1.4.1
Shapely==1.7.0

# authentication packages
flasgger==0.9.4
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
psycopg2--binary==2.8.5
pytest==5.4.3
pytest-cov==2.9.0
flanker==0.9.11
PyJWT==1.7.1
freezegun==0.3.15
